+++ packages/styleguide/.storybook/argTypes/index.tsx
@@ -1,0 +1,2 @@
+export * from './closeButtonProps';
+export * from './infotip';