+++ packages/variance/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.25.0](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.24.1...@codecademy/variance@0.25.0) (2025-10-15)
+
+### Features
+
+- Percipio theme + theme switcher tool ([4307902](https://github.com/Codecademy/gamut/commit/4307902838509a83024781ad70d3bdb77850afdb))
+
+### [0.24.1](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.24.0...@codecademy/variance@0.24.1) (2025-10-14)
+
+**Note:** Version bump only for package @codecademy/variance
+
 ## [0.24.0](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.23.0...@codecademy/variance@0.24.0) (2025-08-20)
 
 ### Features